Extract _QuadForm base and relocate SumSquares
Introduce a private _QuadForm(Atom, ABC) base in atoms/quad_form.py that
holds the shared quadratic-form evaluation (Q=None denotes the identity /
sum of squares) and move SumSquares onto it. Pure refactor with no behavior
change; delete the now-empty sum_squares.py and update the imports.
